Split task implementations into individual files
