+++ handlers/user/userEmailPage.go
@@ -1,64 +1,21 @@
 package user
 
 import (
-	"crypto/rand"
 	"database/sql"
-	"encoding/hex"
-	"errors"
-	"fmt"
-	"github.com/arran4/goa4web/core/consts"
 	"log"
 	"net/http"
 	"net/url"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/arran4/goa4web/core/common"
-	"github.com/arran4/goa4web/internal/tasks"
+	"github.com/arran4/goa4web/core/consts"
 
 	"github.com/arran4/goa4web/handlers"
 	"github.com/arran4/goa4web/internal/middleware"
 
-	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core"
 	"github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/eventbus"
-	notif "github.com/arran4/goa4web/internal/notifications"
 )
-
-type SaveEmailTask struct{ tasks.TaskString }
-
-// AddEmailTask handles user email verification requests and sends
-// notifications directly to the specified address.
-type AddEmailTask struct{ tasks.TaskString }
-
-// ResendVerificationEmailTask resends the verification link for an
-// unverified user email address.
-type ResendVerificationEmailTask struct{ tasks.TaskString }
-type DeleteEmailTask struct{ tasks.TaskString }
-type TestMailTask struct{ tasks.TaskString }
-
-var _ tasks.Task = (*TestMailTask)(nil)
-var _ notif.SelfNotificationTemplateProvider = (*TestMailTask)(nil)
-var _ notif.DirectEmailNotificationTemplateProvider = (*AddEmailTask)(nil)
-var _ notif.DirectEmailNotificationTemplateProvider = (*ResendVerificationEmailTask)(nil)
-
-func (ResendVerificationEmailTask) Action(w http.ResponseWriter, r *http.Request) any {
-	addEmailTask.Resend(w, r)
-	return nil
-}
-
-var (
-	saveEmailTask               = &SaveEmailTask{TaskString: tasks.TaskString(TaskSaveAll)}
-	addEmailTask                = &AddEmailTask{TaskString: tasks.TaskString(TaskAdd)}
-	resendVerificationEmailTask = &ResendVerificationEmailTask{TaskString: TaskResend}
-	deleteEmailTask             = &DeleteEmailTask{TaskString: tasks.TaskString(TaskDelete)}
-	testMailTask                = &TestMailTask{TaskString: tasks.TaskString(TaskTestMail)}
-)
-
-// ErrMailNotConfigured is returned when test mail has no provider configured.
-var ErrMailNotConfigured = errors.New("mail isn't configured")
 
 func userEmailPage(w http.ResponseWriter, r *http.Request) {
 	type Data struct {
@@ -104,260 +61,6 @@ func userEmailPage(w http.ResponseWriter, r *http.Request) {
 	}
 
 	handlers.TemplateHandler(w, r, "emailPage.gohtml", data)
-}
-func (SaveEmailTask) Action(w http.ResponseWriter, r *http.Request) any {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return nil
-	}
-	uid, _ := session.Values["UID"].(int32)
-	if uid == 0 {
-		http.Error(w, "forbidden", http.StatusForbidden)
-		return nil
-	}
-	if err := r.ParseForm(); err != nil {
-		log.Printf("ParseForm Error: %s", err)
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-
-	updates := r.PostFormValue("emailupdates") != ""
-	auto := r.PostFormValue("autosubscribe") != ""
-
-	_, err := cd.Preference()
-	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		log.Printf("preference load: %v", err)
-		http.Redirect(w, r, "?error="+err.Error(), http.StatusTemporaryRedirect)
-		return nil
-	}
-
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			err = queries.InsertEmailPreference(r.Context(), db.InsertEmailPreferenceParams{
-				Emailforumupdates:    sql.NullBool{Bool: updates, Valid: true},
-				AutoSubscribeReplies: auto,
-				UsersIdusers:         uid,
-			})
-		}
-	} else {
-		err = queries.UpdateEmailForumUpdatesByUserID(r.Context(), db.UpdateEmailForumUpdatesByUserIDParams{
-			Emailforumupdates: sql.NullBool{Bool: updates, Valid: true},
-			UsersIdusers:      uid,
-		})
-		if err == nil {
-			err = queries.UpdateAutoSubscribeRepliesByUserID(r.Context(), db.UpdateAutoSubscribeRepliesByUserIDParams{
-				AutoSubscribeReplies: auto,
-				UsersIdusers:         uid,
-			})
-		}
-	}
-	if err != nil {
-		log.Printf("save email pref: %v", err)
-		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-		return nil
-	}
-
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-	return nil
-}
-
-func (TestMailTask) Action(w http.ResponseWriter, r *http.Request) any {
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	user, _ := cd.CurrentUser()
-	if user == nil {
-		http.Error(w, "email unknown", http.StatusBadRequest)
-		return nil
-	}
-	if cd.EmailProvider() == nil {
-		q := url.QueryEscape(ErrMailNotConfigured.Error())
-		r.URL.RawQuery = "error=" + q
-		handlers.TaskErrorAcknowledgementPage(w, r)
-		return nil
-	}
-	if evt := cd.Event(); evt != nil {
-		if evt.Data == nil {
-			evt.Data = map[string]any{}
-		}
-	}
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-	return nil
-}
-
-func (TestMailTask) SelfEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("testEmail")
-}
-
-func (TestMailTask) SelfInternalNotificationTemplate() *string {
-	s := notif.NotificationTemplateFilenameGenerator("testEmail")
-	return &s
-}
-
-func (AddEmailTask) Action(w http.ResponseWriter, r *http.Request) any {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return nil
-	}
-	uid, _ := session.Values["UID"].(int32)
-	if err := r.ParseForm(); err != nil {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return nil
-	}
-	emailAddr := r.FormValue("new_email")
-	if emailAddr == "" {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return nil
-	}
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	if ue, err := queries.GetUserEmailByEmail(r.Context(), emailAddr); err == nil && ue.VerifiedAt.Valid {
-		http.Redirect(w, r, "/usr/email?error=email+exists", http.StatusSeeOther)
-		return nil
-	}
-	var buf [8]byte
-	if _, err := rand.Read(buf[:]); err != nil {
-		log.Printf("rand read: %v", err)
-	}
-	code := hex.EncodeToString(buf[:])
-	expire := time.Now().Add(24 * time.Hour)
-	if err := queries.InsertUserEmail(r.Context(), db.InsertUserEmailParams{UserID: uid, Email: emailAddr, VerifiedAt: sql.NullTime{}, LastVerificationCode: sql.NullString{String: code, Valid: true}, VerificationExpiresAt: sql.NullTime{Time: expire, Valid: true}, NotificationPriority: 0}); err != nil {
-		// TODO better error for NOT Duplicate entry for key 'user_emails_email_code_idx'
-		log.Printf("insert user email: %v", err)
-		http.Redirect(w, r, "/usr/email?error=email+exists", http.StatusSeeOther)
-		return nil
-	}
-	path := "/usr/email/verify?code=" + code
-	page := "http://" + r.Host + path
-	if config.AppRuntimeConfig.HTTPHostname != "" {
-		page = strings.TrimRight(config.AppRuntimeConfig.HTTPHostname, "/") + path
-	}
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	evt := cd.Event()
-	evt.Data["page"] = page
-	evt.Data["email"] = emailAddr
-	evt.Data["URL"] = page
-	if user, err := cd.CurrentUser(); err == nil && user != nil {
-		evt.Data["Username"] = user.Username.String
-	}
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-	return nil
-}
-
-func (AddEmailTask) Resend(w http.ResponseWriter, r *http.Request) {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return
-	}
-	uid, _ := session.Values["UID"].(int32)
-	if err := r.ParseForm(); err != nil {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return
-	}
-	id, _ := strconv.Atoi(r.FormValue("id"))
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	ue, err := queries.GetUserEmailByID(r.Context(), int32(id))
-	if err != nil || ue.UserID != uid {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return
-	}
-	var buf [8]byte
-	if _, err := rand.Read(buf[:]); err != nil {
-		log.Printf("rand read: %v", err)
-	}
-	code := hex.EncodeToString(buf[:])
-	expire := time.Now().Add(24 * time.Hour)
-	if err := queries.SetVerificationCode(r.Context(), db.SetVerificationCodeParams{LastVerificationCode: sql.NullString{String: code, Valid: true}, VerificationExpiresAt: sql.NullTime{Time: expire, Valid: true}, ID: int32(id)}); err != nil {
-		log.Printf("set verification code: %v", err)
-	}
-	path := "/usr/email/verify?code=" + code
-	page := "http://" + r.Host + path
-	if config.AppRuntimeConfig.HTTPHostname != "" {
-		page = strings.TrimRight(config.AppRuntimeConfig.HTTPHostname, "/") + path
-	}
-	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
-	evt := cd.Event()
-	evt.Data["page"] = page
-	evt.Data["email"] = ue.Email
-	evt.Data["URL"] = page
-	if user, err := cd.CurrentUser(); err == nil && user != nil {
-		evt.Data["Username"] = user.Username.String
-	}
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-}
-
-func (DeleteEmailTask) Action(w http.ResponseWriter, r *http.Request) any {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return nil
-	}
-	uid, _ := session.Values["UID"].(int32)
-	if err := r.ParseForm(); err != nil {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return nil
-	}
-	id, _ := strconv.Atoi(r.FormValue("id"))
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	ue, err := queries.GetUserEmailByID(r.Context(), int32(id))
-	if err == nil && ue.UserID == uid {
-		if err := queries.DeleteUserEmail(r.Context(), int32(id)); err != nil {
-			log.Printf("delete user email: %v", err)
-		}
-	}
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-	return nil
-}
-
-func (AddEmailTask) Notify(w http.ResponseWriter, r *http.Request) {
-	session, ok := core.GetSessionOrFail(w, r)
-	if !ok {
-		return
-	}
-	uid, _ := session.Values["UID"].(int32)
-	if err := r.ParseForm(); err != nil {
-		http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-		return
-	}
-	id, _ := strconv.Atoi(r.FormValue("id"))
-	queries := r.Context().Value(consts.KeyCoreData).(*common.CoreData).Queries()
-	val, _ := queries.GetMaxNotificationPriority(r.Context(), uid)
-	var maxPr int32
-	switch v := val.(type) {
-	case int64:
-		maxPr = int32(v)
-	case int32:
-		maxPr = v
-	}
-	if err := queries.SetNotificationPriority(r.Context(), db.SetNotificationPriorityParams{NotificationPriority: maxPr + 1, ID: int32(id)}); err != nil {
-		log.Printf("set notification priority: %v", err)
-	}
-	http.Redirect(w, r, "/usr/email", http.StatusSeeOther)
-}
-
-func (AddEmailTask) DirectEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("verifyEmail")
-}
-
-func (AddEmailTask) DirectEmailAddress(evt eventbus.TaskEvent) (string, error) {
-	if evt.Data != nil {
-		if email, ok := evt.Data["email"].(string); ok {
-			return email, nil
-		}
-	}
-	return "", fmt.Errorf("email not provided")
-}
-
-func (ResendVerificationEmailTask) DirectEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("verifyEmail")
-}
-
-func (ResendVerificationEmailTask) DirectEmailAddress(evt eventbus.TaskEvent) (string, error) {
-	if evt.Data != nil {
-		if email, ok := evt.Data["email"].(string); ok {
-			return email, nil
-		}
-	}
-	return "", fmt.Errorf("email not provided")
 }
 
 func userEmailVerifyCodePage(w http.ResponseWriter, r *http.Request) {
